requirements.txt
chromadb>=0.4.17
ollama>=0.1.5
torch>=2.0.0
transformers>=4.34.0
datasets>=2.14.0
peft>=0.5.0
accelerate>=0.23.0
bitsandbytes>=0.41.0
scipy>=1.11.0
scikit-learn>=1.3.0
numpy>=1.24.0
tqdm>=4.65.0
psutil>=5.9.0
GPUtil>=1.4.0
requests>=2.31.0
logging>=0.5.0
argparse>=1.4.0
json>=2.0.9
typing>=3.7.4
huggingface-hub>=0.19.4
python-dotenv==1.0.0
